requirements.txt
--index-url http://c.pypi.python.org/simple
Django<1.6.0
djangorestframework
south
